PKG -- [util-logger] Rename package to util-logger
