+++ packages/util-logger/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@onflow/logger",
+  "name": "@onflow/util-logger",
   "version": "0.0.1",
   "description": "Logger for FCL-JS",
   "license": "Apache-2.0",
@@ -22,10 +22,10 @@
     "jest-esm-transformer": "1.0.0",
     "microbundle": "0.12.0-next.8"
   },
-  "source": "src/logger.js",
-  "main": "dist/logger.js",
-  "module": "dist/logger.module.js",
-  "unpkg": "dist/logger.umd.js",
+  "source": "src/util-logger.js",
+  "main": "dist/util-logger.js",
+  "module": "dist/util-logger.module.js",
+  "unpkg": "dist/util-logger.umd.js",
   "scripts": {
     "prepublishOnly": "npm test && npm run build",
     "test": "jest",
